@@ -4,7 +4,7 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.18.2 <3.0.0'
 
 dependencies:
   cupertino_icons: ^0.1.2
