@@ -4,7 +4,7 @@ version: 0.0.4
 homepage: https://github.com/grihlo/stacked_card_carousel
 
 environment:
-  sdk: '>=2.12.0-0 <3.0.0'
+  sdk: '>=2.18.2 <3.0.0'
 
 dependencies:
   flutter:
